+++ LSR-Plot-GUI/tsegment.h
@@ -49,21 +49,21 @@ public:
     
     /* getter functions */
     TData* get_dataset_ptr();
-    size_t get_index_from();
-    size_t get_index_to();
+    int get_index_from();
+    int get_index_to();
     double get_ages(int);
     double get_depths(int);
     double get_fit_line(int);
     double get_lsr_plot_value(int);
     double get_smoothed_lsr_plot_value(int);
     double get_lsr_plot_age(int);
-    size_t get_ages_vector_size();
-    size_t get_depths_vector_size();
-    size_t get_fit_line_vector_size();
-    size_t get_lsr_values_vector_size();
-    size_t get_lsr_plot_values_vector_size();
-    size_t get_smoothed_lsr_plot_values_vector_size();
-    size_t get_lsr_plot_ages_vector_size();
+    int get_ages_vector_size();
+    int get_depths_vector_size();
+    int get_fit_line_vector_size();
+    int get_lsr_values_vector_size();
+    int get_lsr_plot_values_vector_size();
+    int get_smoothed_lsr_plot_values_vector_size();
+    int get_lsr_plot_ages_vector_size();
     TGraph* get_g1_ptr();
     TGraph* get_g3_ptr();
